added the new required discount business rules

+++ src/main/scala/etl/loading/DataBaseLoader.scala
@@ -31,7 +31,7 @@ case object DataBaseLoader{
     preparedStatement.setTimestamp(2, stringToTimestamp(order.orderTimestamp)) //order date
     preparedStatement.setString(3, order.productName) // product name
     preparedStatement.setDate(4, stringToDate(order.expiryDate)) // expiry date
-    preparedStatement.setInt(5, order.quantity) // unit price
+    preparedStatement.setDouble(5, order.quantity) // unit price
     preparedStatement.setDouble(6, order.unitPrice) // quantity
     preparedStatement.setDouble(7, order.priceBeforeDiscount) // price before discount
     preparedStatement.setDouble(8, order.discount) // discount
